feat(wealth-library): phase 0 — extract DDReportBody standalone
Extracts the DD Report reading workbench, audit trail, approval
dialogs, SSE generation stream and PDF download into a self-
contained reader component at
lib/components/library/readers/DDReportBody.svelte. Strict prop
contract: only `reportId`. Token comes from the existing
"netz:getToken" Svelte context; optional "netz:dd-actor" context
exposes actor info for the approval UI so the future
LibraryPreviewPane can render the same body in a read-only host
without ever showing approve/reject buttons.

Toxic toFixed() on confidence_score replaced with formatPercent
from @investintell/ui per CLAUDE.md formatter discipline.

The legacy route at screener/dd-reports/[fundId]/[reportId]
becomes a thin shell: PageHeader (title + breadcrumbs) plus
<DDReportBody {reportId} /> under the route data branches. The
server loader is unchanged so backend contract is preserved.
